# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CyberArk Identity Sample NodeJS Server",
        "version": "1.0.0",
        "title": "NodeJS API",
        "contact": {
            "email": "abc@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": ["http"],
    "host": "localhost:2200",
    "basePath": "/api",
    "paths" : {
        "/auth/beginAuth" : {
            "post" : {
                "tags": [
                    "Auth"
                ],
                "summary" : "Begin's authentication",
                "description": "Begin's authentication",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "state": {
                                    "type": "string"
                                },
                                "mechanisms": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "AnswerType": {
                                                "type": "string"
                                            },
                                            "Name": {
                                                "type": "string"
                                            },
                                            "PromptMechChosen": {
                                                "type": "string"
                                            },
                                            "PromptSelectMech": {
                                                "type": "string"
                                            },
                                            "MechanismId": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/advanceAuth" : {
            "post" : {
                "tags": [
                    "Auth"
                ],
                "summary" : "For simple text answer mechanisms",
                "description": "For simple text answer mechanisms",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "mechanismId" : {
                                    "type": "string"
                                },
                                "answer" : {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "state": {
                                    "type": "string"
                                },
                                "mechanisms": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                },
                                "userInfo": {
                                    "type": "object",
                                    "properties": {
                                        "email": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "authToken": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}